requirements.txt
astroid==2.11.5
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
colorama==0.4.4
dill==0.3.4
distlib==0.3.4
filelock==3.6.0
identify==2.5.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mysql-connector-python==8.0.29
nodeenv==1.6.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.19.0
protobuf==3.20.1
py==1.11.0
pylint==2.13.8
pyparsing==3.0.9
pytest==7.1.2
PyYAML==6.0
requests==2.27.1
requests-mock==1.9.3
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
wrapt==1.14.1
